# docs/swagger.yaml
swagger: '2.0'
info:
  title: Living Progress - Build - Mentor Me
  description: Living Progress - Build - Mentor Me API.
  version: "1.0.0"
# the sample made up domain of the service
host: api.mentorme.com
...
schemes:
  - http
  - https
# will be prefixed to all paths
securityDefinitions:
  basicAuth:
    type: basic
# HTTP Basic Authentication. Works over `HTTP` and `HTTPS`
basePath: /v1
produces:
  - application/json
paths:
  /institutions/{id}:
    get:
      tags:
        - Institution
      summary: get institution
      description: This endpoint is responsible for getting institution
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the Institutions
          schema:
            $ref: '#/definitions/Institution'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - ien is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    put:
      tags:
        - Institution
      summary: update institution
      description: This endpoint is responsible for updating institution
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
        - name: body
          in: body
          description: The institution to update
          schema:
            $ref: '#/definitions/Institution'
      responses:
        200:
          description: The response containing the Institution
          schema:
            $ref: '#/definitions/Institution'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'

    delete:
      tags:
        - Institution
      summary: delte institution
      description: This endpoint is responsible for deleting institution
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /institutions:
    get:
          tags:
            - Institution
          summary: search for entities
          description: This endpoint is responsible for searching by criteria. For the rest of the criteria params see TCUML.
          security:
          - basicAuth: []
          parameters:
            - name: institutionName
              in: query
              description: The institution Name
              type: string
              required: false
            - name: parentOrganization
              in: query
              description: The parent organization
              type: string
              required: false
            - name: location
              in: query
              description: The location
              type: string
              required: false
            - name: email
              in: query
              description: The email
              type: string
              required: false
            - name: description
              in: query
              description: The description
              type: string
              required: false
            - name: status
              in: query
              description: The status
              type: string
              required: false
            - name: pageNumber
              in: query
              description: The page number
              type: number
              format: int32
              required: false
            - name: pageSize
              in: query
              description: The page size
              type: number
              format: int32
              required: false
            - name: sortColumn
              in: query
              description: The sort column
              type: string
              required: false
            - name: sortOrder
              in: query
              description: The sort order
              type: string
              required: false
          responses:
            200:
              description: The response containing the Institutions
              schema:
                $ref: '#/definitions/InstitutionSearchResult'
            400:
              description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
              schema:
                $ref: '#/definitions/Error'
            401:
              description: Authentication Error - if token is missing or invalid
              schema:
                $ref: '#/definitions/Error'
            403:
              description: Access Denied Error - if forbidden to access
              schema:
                $ref: '#/definitions/Error'
            500:
              description: Internal server error - if the request was properly formatted, but the operation failed on the server side
              schema:
                $ref: '#/definitions/Error'
            default:
              description: Unexpected error.
              schema:
                $ref: '#/definitions/Error'
    post:
      tags:
        - Institution
      summary: create institution
      description: This endpoint is responsible for creating institution
      security:
        - basicAuth: []
      parameters:
        - name: institution
          in: body
          description: The institution to create
          schema:
            $ref: '#/definitions/InstitutionCreateBody'
      responses:
        200:
          description: The response containing the Institution
          schema:
            $ref: '#/definitions/Institution'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
            description: Authentication Error - if token is missing or invalid
            schema:
              $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /institutions/{id}/generateAffiliationCode:
    put:
      tags:
        - Institution
      summary: get generated institution affiliation code
      description: This endpoint is responsible for getting institution affiliation code
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the code
          schema:
            type: string
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /institutions/affiliationCode/{affiliationCode}:
    get:
      tags:
        - Institution
      summary: get  institution for affiliation code
      description: This endpoint is responsible for getting institution for affiliation code
      security:
        - basicAuth: []
      parameters:
        - name: affiliationCode
          in: path
          description: The affiliation code
          type: string
          required: true
      responses:
        200:
          description: The response containing the Institution
          schema:
            $ref: '#/definitions/Institution'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /institutions/useAffiliationCode:
    put:
      tags:
        - Institution
      summary: use affiliation code
      description: This endpoint is responsible for setting used flag for affiliation code
      security:
        - basicAuth: []
      parameters:
        - name: affiliationCode
          in: query
          description: The affiliation code
          type: string
          required: true
      responses:
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
  /institutions/{id}/summary:
    get:
      tags:
        - Institution
      summary: get the summary
      description: This endpoint is responsible for getting institution summary
      security:
        - basicAuth: []
      parameters:
        - name: id
          in: path
          description: The id
          type: string
          required: true
      responses:
        200:
          description: The response containing the InstitutionSummary
          schema:
            $ref: '#/definitions/InstitutionSummary'
        400:
          description: Bad Request - if there was problem with the request (e.g. malformed or some parameters are missing)
          schema:
            $ref: '#/definitions/Error'
        401:
          description: Authentication Error - if token is missing or invalid
          schema:
            $ref: '#/definitions/Error'
        403:
          description: Access Denied Error - if forbidden to access
          schema:
            $ref: '#/definitions/Error'
        404:
          description: Not Found - if match entity not found
          schema:
            $ref: '#/definitions/Error'
        500:
          description: Internal server error - if the request was properly formatted, but the operation failed on the server side
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error.
          schema:
            $ref: '#/definitions/Error'
...
definitions:
...
  Country:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
  State:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      value:
        type: string
        description: The value.
...
  InstitutionContact:
      type: object
      properties:
        id:
          type: integer
          format: int32
          description: The id.
        title:
          type: string
          description: The title.
        firstName:
          type: string
          description: The first name.
        lastName:
          type: string
          description: The last name.
        email:
          type: string
          description: The email.
        phoneNumber:
          type: string
          description: The phone number.
        primaryContact:
          type: boolean
          description: The primary contact flag.
...
  InstitutionCreateBody:
    type: object
    properties:
      institutionName:
        type: string
        description: The name.
      parentOrganization:
        type: string
        description: The parent organization.
      streetAddress:
        type: string
        description: The street address.
      city:
        type: string
        description: The city.
      zip:
        type: string
        description: The zip.
      phone:
         type: string
         description: The phone.
      email:
        type: string
        description: The email.
      description:
        type: string
        description: The description.
      status:
        type: string
        description: The status.
      logoPath:
        type: string
        description: The logo path.
      contacts:
        type: array
        items:
          $ref: '#/definitions/InstitutionContact'
      state:
        $ref: '#/definitions/State'
      country:
        $ref: '#/definitions/Country'
  Institution:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: The id.
      createdOn:
        type: integer
        format: int32
        description: The created on datetime with unix epoch format.
      institutionName:
        type: string
        description: The name.
      parentOrganization:
        type: string
        description: The parent organization.
      streetAddress:
        type: string
        description: The street address.
      city:
        type: string
        description: The city.
      zip:
        type: string
        description: The zip.
      phone:
         type: string
         description: The phone.
      email:
        type: string
        description: The email.
      description:
        type: string
        description: The description.
      status:
        type: string
        description: The status.
      logoPath:
        type: string
        description: The logo path.
      contacts:
        type: array
        items:
          $ref: '#/definitions/InstitutionContact'
      state:
        $ref: '#/definitions/State'
      country:
        $ref: '#/definitions/Country'
  InstitutionSummary:
    type: object
    properties:
      institutionalProgramsCount:
        type: number
        format: integer
        description: The institutional programs count.
      upcomingEventsCount:
        type: number
        format: integer
        description: The events count.
      associatedMenteesCount:
        type: number
        format: integer
        description: The mentees count.
      associatedMentorsCount:
        type: number
        format: integer
        description: The mentors count.
...
  InstitutionSearchResult:
    type: object
    properties:
      total:
        type: number
        format: integer
        description: The total.
      totalPages:
        type: number
        format: integer
        description: The total pages.
      entities:
        type: array
        items:
          $ref: '#/definitions/Institution'
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string